refactoring and rearrangement to comply with merge
