+++ components/Navbar/Navbar.js
@@ -1,12 +1,11 @@
 import Image from "next/image";
-
-
+import Link from "next/link";
 const Navbar = () => {
   return (
     <nav className="navbar navbar-expand-lg  navbar-dark bg-dark ">
       <div className="container-fluid">
         <a className="navbar-brand" href="#">
-          <Image src="/logo.png" width={130} height={77} alt="" />{" "}
+          <Image src="/assets/images/logo.png" width={130} height={77} alt="" />{" "}
         </a>
         <button
           className="navbar-toggler"
@@ -22,9 +21,11 @@ const Navbar = () => {
         <div className="collapse navbar-collapse" id="navbarNavDropdown">
           <ul className="navbar-nav ms-auto">
             <li className="nav-item me-5">
+              <Link href={"/"}>
               <a className="nav-link" aria-current="page" href="#">
                 Home
               </a>
+              </Link>
             </li>
             <li className="nav-item me-5 ">
               <a className="nav-link" aria-current="page" href="#">
@@ -48,9 +49,9 @@ const Navbar = () => {
                 aria-labelledby="navbarDropdownMenuLink"
               >
                 <li>
-                  <a className="dropdown-item" href="#">
-                    Membership Drive
-                  </a>
+                  <Link href="/membership-drive">
+                    <a className="dropdown-item">Membership Drive</a>
+                  </Link>
                 </li>
                 <li>
                   <a className="dropdown-item" href="#">
